@@ -5,7 +5,7 @@ author: Joshua Lorincz <joshlordeveloper@gmail.com>
 homepage: https://github.com/JoshLorDeveloper/currency_pickers
 
 environment:
-  sdk: ">=2.0.0-dev.68.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
