@@ -14,28 +14,28 @@ class CurrencyPickerDropdown extends StatefulWidget {
   });
 
   /// Filters the available country list
-  final ItemFilter itemFilter;
+  final ItemFilter? itemFilter;
 
   ///This function will be called to build the child of DropdownMenuItem
   ///If it is not provided, default one will be used which displays
   ///flag image, isoCode and currencyCode in a row.
   ///Check _buildDefaultMenuItem method for details.
-  final ItemBuilder itemBuilder;
+  final ItemBuilder? itemBuilder;
 
   ///It should be one of the ISO ALPHA-2 Code that is provided
   ///in countryList map of countries.dart file.
-  final String initialValue;
+  final String? initialValue;
 
   ///This function will be called whenever a Country item is selected.
-  final ValueChanged<Country> onValuePicked;
+  final ValueChanged<Country?>? onValuePicked;
 
   @override
   _CurrencyPickerDropdownState createState() => _CurrencyPickerDropdownState();
 }
 
 class _CurrencyPickerDropdownState extends State<CurrencyPickerDropdown> {
-  List<Country> _countries;
-  Country _selectedCountry;
+  List<Country> _countries = [];
+  Country? _selectedCountry;
 
   @override
   void initState() {
@@ -46,7 +46,7 @@ class _CurrencyPickerDropdownState extends State<CurrencyPickerDropdown> {
     if (widget.initialValue != null) {
       try {
         _selectedCountry = _countries.firstWhere(
-          (country) => country.isoCode == widget.initialValue.toUpperCase(),
+          (country) => country.isoCode == widget.initialValue?.toUpperCase(),
         );
       } catch (error) {
         throw Exception(
@@ -65,7 +65,7 @@ class _CurrencyPickerDropdownState extends State<CurrencyPickerDropdown> {
         .map((country) => DropdownMenuItem<Country>(
             value: country,
             child: widget.itemBuilder != null
-                ? widget.itemBuilder(country)
+                ? widget.itemBuilder?.call(country) ?? SizedBox()
                 : _buildDefaultMenuItem(country)))
         .toList();
 
@@ -77,7 +77,7 @@ class _CurrencyPickerDropdownState extends State<CurrencyPickerDropdown> {
             onChanged: (value) {
               setState(() {
                 _selectedCountry = value;
-                widget.onValuePicked(value);
+                widget.onValuePicked?.call(value);
               });
             },
             items: items,
